2021/1/14 调拨记录完成

+++ xboot/xboot-plus-back-modules/xboot-modules/xboot-zpig/src/main/java/serviceimpl/CollectionRecordServiceImpl.java
@@ -2,6 +2,7 @@ package serviceimpl;
 
 import com.baomidou.dynamic.datasource.annotation.DS;
 import dao.CtPigCollectionrecordDao;
+import dao.CtPigSemencollectionDao;
 import entity.CtPigCollectionrecord;
 import entity.TPmUser;
 import lombok.extern.slf4j.Slf4j;
@@ -23,10 +24,11 @@ import java.util.List;
 @DS("slave1")
 public class CollectionRecordServiceImpl implements CollectionRecordService {
     private final CtPigCollectionrecordDao ctPigCollectionrecordDao;
-
+    private final CtPigSemencollectionDao ctPigSemencollectionDao;
     @Autowired
-    public CollectionRecordServiceImpl(CtPigCollectionrecordDao ctPigCollectionrecordDao, OkHttpClient okHttpClient) {
+    public CollectionRecordServiceImpl(CtPigCollectionrecordDao ctPigCollectionrecordDao, OkHttpClient okHttpClient, CtPigSemencollectionDao ctPigSemencollectionDao) {
         this.ctPigCollectionrecordDao = ctPigCollectionrecordDao;
+        this.ctPigSemencollectionDao = ctPigSemencollectionDao;
     }
 
     @Override
@@ -41,7 +43,13 @@ public class CollectionRecordServiceImpl implements CollectionRecordService {
 
     @Override
     public int deleteCollectionRecord(String fId) {
-        return ctPigCollectionrecordDao.deleteByPrimaryKey(fId);
+        CtPigCollectionrecord ctPigCollectionrecord = ctPigCollectionrecordDao.selectByPrimaryKey(fId);
+        int audiType = ctPigCollectionrecord.getCfifauditing().intValue();
+        if (audiType == 1){
+            return 0;
+        }else {
+            return ctPigCollectionrecordDao.deleteByPrimaryKey(fId);
+        }
     }
 
     @Override
@@ -72,6 +80,9 @@ public class CollectionRecordServiceImpl implements CollectionRecordService {
         String userId = tPmUser.getFid();
         for (CtPigCollectionrecord ctPigCollectionrecord : records) {
             String fid = ctPigCollectionrecord.getFid();
+            String appFid = ctPigCollectionrecord.getCfdesignnumberid();
+            //审核时更新精液采集计划状态
+            ctPigSemencollectionDao.updatDesign(appFid);
             results.add(ctPigCollectionrecordDao.updateIfAuditByFid(fid , userId));
         }
         return results;
